requirements.txt
fastapi
uvicorn[standard]
yt-dlp
jinja2
python-multipart
certifi
openai
python-dotenv
requests
python-telegram-bot==20.8
